// apps/server/src/generated/prisma/models/AiPromptConfig.ts
export type AiPromptConfigModel = runtime.Types.Result.DefaultSelection<Prisma.$AiPromptConfigPayload>;
// ...
export type $AiPromptConfigPayload<
  ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs
> = {
  name: 'AiPromptConfig';
  objects: {};
  scalars: runtime.Types.Extensions.GetPayloadResult<
    {
      id: string;
      promptKey: string;
      title: string;
      systemPrompt: string;
      createdAt: Date;
      updatedAt: Date;
    },
    ExtArgs['result']['aiPromptConfig']
  >;
  composites: {};
};